[쇼핑 주문 3, 4단계] 크림 미션 제출합니다. (#140)
* refactor: ID를 Int -> Long 자료형으로 수정

* refactor: CatalogRemoteDataSource 추상화 및 Repository와 분리

* refactor: CartRemoteDataSource 추상화 및 Repository와 분리

* refactor: 카탈로그 Repository, DataSource 코루틴 적용

* refactor: 장바구니 Repository, DataSource, CatalogViewModel 코루틴 적용

* refactor: 최근 본 상품 관련 Repository, ViewModel 비동기 처리 콜백 함수에서 코루틴으로 변경

- RecentlyViewedProductDao의 함수들을 suspend 함수로 변경
- RecentlyViewedProductRepository의 함수들을 suspend 함수로 변경
- DetailViewModel의 최근 본 상품 저장, 불러오기 로직을 코루틴으로 변경
- CartViewModel의 추천 상품 불러오기 로직을 코루틴으로 변경
- CatalogViewModel의 최근 본 상품 불러오기 로직을 코루틴으로 변경

* test: JUnit5 CoroutinesTestExtension 추가

* build: MockK 의존성 추가

* test: CartViewModelTest 코루틴 테스트로 변경 및 Mock 객체 사용

* test: CatalogViewModelTest 코루틴 테스트로 변경 및 Mock 객체 사용

* test: DetailViewModelTest 코루틴 테스트로 변경 및 Mock 객체 사용

* fix: CartItemService suspend 함수로 변경 및 RemoteCartProductRepositoryImpl 누락된 함수 추가

* feat: 주문하기 클릭 시 장바구니 추천상품 Fragment로 전환하는 기능 구현

* fix: 장바구니 아이템 삽입 후 CartItemId 반환토록 수정

* feat: 장바구니 상품 수량 1개에서 감소 시 삭제 기능 추가

* feat: 추천 Fragment 장바구니에 이미 추가된 상품이라면 미노출하도록 구현

* chore: 사용하지 않는 코드 제거

* feat: 주문 관련 DTO 클래스 추가

* feat: 주문에 필요한 Coupon, AvailableTime Domain 모델 추가

* feat: OrderActivity 및 레이아웃 추가

* feat: CouponService, CouponDataSource, CouponRepository 구현

* feat: 쿠폰 RecyclerView 적용

* feat: OrderingProducts 일급 컬렉션 추가 및 주문 화면에 상품 목록 전달 기능 추가

* feat: sealed class CouponType 추가

* feat: OrderViewModel 생성자에 상품 전달하도록 구현

* feat: OrderingProducts 모델 할인 금액 계산 로직 구현 및 레이아웃 ViewModel 연결

* feat: 쿠폰 적용 기능 구현

- 쿠폰 선택 시 할인 금액이 변경되도록 구현
- 쿠폰을 다시 선택하면 선택이 해제되고 할인 금액이 원복되도록 구현
- 정률 할인 쿠폰의 할인 금액이 정상적으로 계산되도록 수정

* refactor: 장바구니에서 주문 시 체크된 상품만 전달하도록 수정

* feat: 쿠폰 선택 시 체크 표시 기능 추가 및 단일 선택되도록 수정

* docs: README 기능 요구 사항 업데이트

* refactor: 확장 함수 `parcelableArray`를 `IntentCompat.kt`로 이동

* feat: 주문 기능 구현 및 주문 완료 시 상품 목록으로 이동

- OrderActivity
    - 주문 완료 시 상품 목록(CatalogActivity)으로 이동하고 기존 스택을 모두 제거하도록 `startActivityWhenOrderMade()` 메서드 추가
    - 주문하기 버튼 클릭 시 `OrderViewModel.makeOrder()` 호출
- OrderViewModel
    - `makeOrder()` 메서드 추가: `OrderRepository.insertOrders()`를 호출하여 주문을 시도하고, 결과를 `_isOrderMade` LiveData에 반영
- OrderRepository, OrderRepositoryImpl 추가
    - 주문 생성을 위한 `insertOrders()` 메서드 정의 및 구현
- OrderService 추가
    - 주문 생성을 위한 `/orders` POST API 정의
- Orders DTO 추가
    - 주문 생성 요청 시 cartItemIds를 담기 위한 DTO
- DataSourceProvider, RepositoryProvider
    - OrderService, OrderRepository 의존성 주입 설정
- OrderViewModelFactory
    - OrderRepository 의존성 주입 설정
- ProductAdapter
    - 불필요한 로그 제거

* refactor: ProductAdapter에 DiffUtil 적용 및 ListAdapter로 변경

* feat: 주문 상품에 쿠폰 만료일, 사용 가능 시간 적용

* docs: README 요구 사항 업데이트

* refactor: DependencyProvider를 DatabaseProvider로 이름 변경

* refactor: CartActivity에서 프래그먼트 교체 로직을 메서드로 분리

* refactor: CartSelectionFragment에서 `loadSelectAllState()` 호출 제거

불필요한 `loadSelectAllState()` 호출을 제거하여 코드 정리

* fix: 장바구니 상품 체크 여부 반영

* feat: 주문할 상품이 없을 경우 주문 버튼 비활성화

* build: kotlinx-coroutines-android, mockk 버전 라이브러리 카탈로그로 관리

* refactor: localhost 도메인 제거 및 usesCleartextTraffic 속성 제거

* refactor: viewModelScope 내 LiveData postValue 대신 setValue 사용

* fix: CouponService 변수명 오타 수정

* refactor: 네트워크 통신 결과 Result 타입으로 변경 및 에러 처리 추가

- Repository 및 ViewModel에서 네트워크 통신 결과를 Result 타입으로 변경했습니다.
- 통신 실패 시 사용자에게 Toast 메시지를 통해 에러를 알리는 기능을 추가했습니다.

* refactor: Retrofit Service Response 반환값 제거 및 CouponRepository, OrderRepository에 Result 타입으로 변경 및 에러 처리 추가

* refactor: DataSource 코드 withContext 제거

Retrofit의 Call Adapter는 이미 IO 스레드에서 네트워크 요청을 처리하므로 withContext(Dispatchers.IO)를 제거합니다.

* fix: 최근 본 상품이 없을 때, 최근 본 상품을 로드할 경우 발생하는 버그 수정

* refactor: Coupons를 List<Coupon>으로 변경

* refactor: CartViewModel.orderClicked LiveData의 값으로 주문할 상품 목록을 전달

* refactor: 주문하기 버튼 클릭 시 프래그먼트 상태 비교 로직 수정

- 기존의 fragmentState 변수를 사용한 상태 비교에서, 현재 표시되고 있는 프래그먼트의 타입을 직접 확인하는 방식으로 변경했습니다.
- 이는 프래그먼트 상태 관리의 명확성을 높이고 잠재적인 오류를 줄이기 위함입니다.

+++ app/src/main/java/woowacourse/shopping/data/entity/CartProductEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "CartProducts")
 data class CartProductEntity(
-    @PrimaryKey(autoGenerate = true) val uid: Int,
+    @PrimaryKey(autoGenerate = true) val uid: Long,
     val imageUrl: String,
     val name: String,
     val price: Int,
